[32-bit] Fix build with uintptr and -O0 https://bugs.webkit.org/show_bug.cgi?id=311133
Reviewed by Yusuke Suzuki.

This fixes the 32-bit armv7 build with a few improvements:

- We detect armv7 containers in more cases (they report armv8l)
- We use O0 for some big files that cause memory exhaustion when not
  cross-compiling. These files could not be compiled before, so this is
  not a regression.
- We update some places in libpas so that it can build on 32-bit,
  in preparation for some future code sharing. It does not need to run.

* Source/WTF/wtf/glib/RunLoopGLib.cpp:
(WTF::RunLoop::RunLoop):
* Source/WebCore/CMakeLists.txt:
* Source/WebCore/loader/ResourceMonitorChecker.cpp:
(WebCore::ResourceMonitorChecker::finishPendingQueries):
* Source/WebKit/CMakeLists.txt:
* Source/WebKit/GPUProcess/GPUProcess.cpp:
(WebKit::GPUProcess::updateSandboxAccess):
* Source/WebKit/NetworkProcess/NetworkResourceLoader.cpp:
(WebKit::NetworkResourceLoader::logSlowCacheRetrieveIfNeeded):
* Source/WebKit/Platform/IPC/Decoder.cpp:
(IPC::copyBuffer):
(IPC::Decoder::create):
* Source/WebKit/UIProcess/Extensions/WebExtension.cpp:
(WebKit::WebExtension::populateDeclarativeNetRequestPropertiesIfNeeded):
* Source/WebKit/UIProcess/WebProcessProxy.cpp:
(WebKit::WebProcessProxy::canTerminateAuxiliaryProcess):
* Source/bmalloc/libpas/src/libpas/pas_utils.c:
(pas_crash_with_info_impl):
(pas_panic):
(pas_assertion_failed_no_inline):
(pas_assertion_failed_no_inline_with_extra_detail):
* Source/bmalloc/libpas/src/libpas/pas_utils.h:
(pas_compare_and_swap_uintptr_weak):
(pas_compare_and_swap_uintptr_strong):
(pas_compare_and_swap_ptr_weak):
(pas_compare_and_swap_ptr_strong):
(pas_compare_and_swap_pair_weak):
(pas_compare_and_swap_pair_strong):
(pas_atomic_load_pair_relaxed):
(pas_atomic_store_pair):
(pas_atomic_store_pair_relaxed):
* Source/cmake/WebKitCommon.cmake:
* Source/cmake/WebKitCompilerFlags.cmake:
* Tools/Scripts/webkitdirs.pm:
(generateBuildSystemFromCMakeProject):

Canonical link: https://commits.webkit.org/310466@main

+++ Source/bmalloc/libpas/src/libpas/pas_utils.c
@@ -85,72 +85,72 @@
 extern "C" {
 #endif
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl0(uint64_t reason)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl0(size_t reason)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR));
     __builtin_unreachable();
 }
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl1(uint64_t reason, uint64_t misc1)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl1(size_t reason, size_t misc1)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR));
     __builtin_unreachable();
 }
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl2(uint64_t reason, uint64_t misc1, uint64_t misc2)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl2(size_t reason, size_t misc1, size_t misc2)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
-    register uint64_t misc2GPR __asm__(CRASH_GPR2) = misc2;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t misc2GPR __asm__(CRASH_GPR2) = misc2;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR), "r"(misc2GPR));
     __builtin_unreachable();
 }
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl3(uint64_t reason, uint64_t misc1, uint64_t misc2, uint64_t misc3)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl3(size_t reason, size_t misc1, size_t misc2, size_t misc3)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
-    register uint64_t misc2GPR __asm__(CRASH_GPR2) = misc2;
-    register uint64_t misc3GPR __asm__(CRASH_GPR3) = misc3;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t misc2GPR __asm__(CRASH_GPR2) = misc2;
+    register size_t misc3GPR __asm__(CRASH_GPR3) = misc3;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR), "r"(misc2GPR), "r"(misc3GPR));
     __builtin_unreachable();
 }
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl4(uint64_t reason, uint64_t misc1, uint64_t misc2, uint64_t misc3, uint64_t misc4)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl4(size_t reason, size_t misc1, size_t misc2, size_t misc3, size_t misc4)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
-    register uint64_t misc2GPR __asm__(CRASH_GPR2) = misc2;
-    register uint64_t misc3GPR __asm__(CRASH_GPR3) = misc3;
-    register uint64_t misc4GPR __asm__(CRASH_GPR4) = misc4;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t misc2GPR __asm__(CRASH_GPR2) = misc2;
+    register size_t misc3GPR __asm__(CRASH_GPR3) = misc3;
+    register size_t misc4GPR __asm__(CRASH_GPR4) = misc4;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR), "r"(misc2GPR), "r"(misc3GPR), "r"(misc4GPR));
     __builtin_unreachable();
 }
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl5(uint64_t reason, uint64_t misc1, uint64_t misc2, uint64_t misc3, uint64_t misc4, uint64_t misc5)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl5(size_t reason, size_t misc1, size_t misc2, size_t misc3, size_t misc4, size_t misc5)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
-    register uint64_t misc2GPR __asm__(CRASH_GPR2) = misc2;
-    register uint64_t misc3GPR __asm__(CRASH_GPR3) = misc3;
-    register uint64_t misc4GPR __asm__(CRASH_GPR4) = misc4;
-    register uint64_t misc5GPR __asm__(CRASH_GPR5) = misc5;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t misc2GPR __asm__(CRASH_GPR2) = misc2;
+    register size_t misc3GPR __asm__(CRASH_GPR3) = misc3;
+    register size_t misc4GPR __asm__(CRASH_GPR4) = misc4;
+    register size_t misc5GPR __asm__(CRASH_GPR5) = misc5;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR), "r"(misc2GPR), "r"(misc3GPR), "r"(misc4GPR), "r"(misc5GPR));
     __builtin_unreachable();
 }
 
-PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl6(uint64_t reason, uint64_t misc1, uint64_t misc2, uint64_t misc3, uint64_t misc4, uint64_t misc5, uint64_t misc6)
+PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl6(size_t reason, size_t misc1, size_t misc2, size_t misc3, size_t misc4, size_t misc5, size_t misc6)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
-    register uint64_t misc2GPR __asm__(CRASH_GPR2) = misc2;
-    register uint64_t misc3GPR __asm__(CRASH_GPR3) = misc3;
-    register uint64_t misc4GPR __asm__(CRASH_GPR4) = misc4;
-    register uint64_t misc5GPR __asm__(CRASH_GPR5) = misc5;
-    register uint64_t misc6GPR __asm__(CRASH_GPR6) = misc6;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t misc2GPR __asm__(CRASH_GPR2) = misc2;
+    register size_t misc3GPR __asm__(CRASH_GPR3) = misc3;
+    register size_t misc4GPR __asm__(CRASH_GPR4) = misc4;
+    register size_t misc5GPR __asm__(CRASH_GPR5) = misc5;
+    register size_t misc6GPR __asm__(CRASH_GPR6) = misc6;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR), "r"(misc2GPR), "r"(misc3GPR), "r"(misc4GPR), "r"(misc5GPR), "r"(misc6GPR));
     __builtin_unreachable();
 }
@@ -162,15 +162,15 @@ PAS_NEVER_INLINE PAS_NO_RETURN void pas_crash_with_info_impl6(uint64_t reason, u
 
 #endif /* PAS_OS(DARWIN) && PAS_VA_OPT_SUPPORTED */
 
-PAS_NEVER_INLINE PAS_NO_RETURN static void pas_crash_with_info_impl(uint64_t reason, uint64_t misc1, uint64_t misc2, uint64_t misc3, uint64_t misc4, uint64_t misc5, uint64_t misc6)
+PAS_NEVER_INLINE PAS_NO_RETURN static void pas_crash_with_info_impl(size_t reason, size_t misc1, size_t misc2, size_t misc3, size_t misc4, size_t misc5, size_t misc6)
 {
-    register uint64_t reasonGPR __asm__(CRASH_GPR0) = reason;
-    register uint64_t misc1GPR __asm__(CRASH_GPR1) = misc1;
-    register uint64_t misc2GPR __asm__(CRASH_GPR2) = misc2;
-    register uint64_t misc3GPR __asm__(CRASH_GPR3) = misc3;
-    register uint64_t misc4GPR __asm__(CRASH_GPR4) = misc4;
-    register uint64_t misc5GPR __asm__(CRASH_GPR5) = misc5;
-    register uint64_t misc6GPR __asm__(CRASH_GPR6) = misc6;
+    register size_t reasonGPR __asm__(CRASH_GPR0) = reason;
+    register size_t misc1GPR __asm__(CRASH_GPR1) = misc1;
+    register size_t misc2GPR __asm__(CRASH_GPR2) = misc2;
+    register size_t misc3GPR __asm__(CRASH_GPR3) = misc3;
+    register size_t misc4GPR __asm__(CRASH_GPR4) = misc4;
+    register size_t misc5GPR __asm__(CRASH_GPR5) = misc5;
+    register size_t misc6GPR __asm__(CRASH_GPR6) = misc6;
     __asm__ volatile (PAS_FATAL_CRASH_INST : : "r"(reasonGPR), "r"(misc1GPR), "r"(misc2GPR), "r"(misc3GPR), "r"(misc4GPR), "r"(misc5GPR), "r"(misc6GPR));
     __builtin_trap();
 }
@@ -178,7 +178,7 @@ PAS_NEVER_INLINE PAS_NO_RETURN static void pas_crash_with_info_impl(uint64_t rea
 #else
 
 PAS_IGNORE_WARNINGS_BEGIN("unused-parameter")
-PAS_NEVER_INLINE PAS_NO_RETURN static void pas_crash_with_info_impl(uint64_t reason, uint64_t misc1, uint64_t misc2, uint64_t misc3, uint64_t misc4, uint64_t misc5, uint64_t misc6) { __builtin_trap(); }
+PAS_NEVER_INLINE PAS_NO_RETURN static void pas_crash_with_info_impl(size_t reason, size_t misc1, size_t misc2, size_t misc3, size_t misc4, size_t misc5, size_t misc6) { __builtin_trap(); }
 PAS_IGNORE_WARNINGS_END
 
 #endif
@@ -191,7 +191,7 @@ void pas_panic(const char* format, ...)
         pas_log("[%d] pas panic: ", getpid());
         va_start(arg_list, format);
         pas_vlog(format, arg_list);
-        pas_crash_with_info_impl((uint64_t)format, 0, 0, 0, 0, 0, 0);
+        pas_crash_with_info_impl((size_t)format, 0, 0, 0, 0, 0, 0);
     }
     __builtin_trap();
 }
@@ -209,14 +209,14 @@ void pas_assertion_failed_no_inline(const char* filename, int line, const char* 
 {
     pas_log("[%d] pas assertion failed: ", getpid());
     pas_log("%s:%d: %s: assertion %s failed.\n", filename, line, function, expression);
-    pas_crash_with_info_impl((uint64_t)filename, (uint64_t)line, (uint64_t)function, (uint64_t)expression, 0xbeefbff0, 42, 1337);
+    pas_crash_with_info_impl((size_t)filename, (size_t)line, (size_t)function, (size_t)expression, 0xbeefbff0, 42, 1337);
 }
 
-void pas_assertion_failed_no_inline_with_extra_detail(const char* filename, int line, const char* function, const char* expression, uint64_t extra)
+void pas_assertion_failed_no_inline_with_extra_detail(const char* filename, int line, const char* function, const char* expression, size_t extra)
 {
     pas_log("[%d] pas assertion failed (with extra detail): ", getpid());
-    pas_log("%s:%d: %s: assertion %s failed. Extra data: %" PRIu64 ".\n", filename, line, function, expression, extra);
-    pas_crash_with_info_impl((uint64_t)filename, (uint64_t)line, (uint64_t)function, (uint64_t)expression, extra, 1337, 0xbeef0bff);
+    pas_log("%s:%d: %s: assertion %s failed. Extra data: %" PRIuPTR ".\n", filename, line, function, expression, extra);
+    pas_crash_with_info_impl((size_t)filename, (size_t)line, (size_t)function, (size_t)expression, extra, 1337, 0xbeef0bff);
 }
 
 void pas_panic_on_out_of_memory_error(void)
